# Buildcache tarballs with rootfs structure