--- a/lib/spack/spack/test/bindist.py
+++ b/lib/spack/spack/test/bindist.py
@@ -899,22 +899,21 @@ def test_tarball_doesnt_include_buildinfo_twice(tmpdir):
     tarball = str(tmpdir.join("prefix.tar.gz"))
 
     bindist._do_create_tarball(
-        tarfile_path=tarball,
-        binaries_dir=str(p),
-        pkg_dir="my-pkg-prefix",
-        buildinfo={"metadata": "new"},
+        tarfile_path=tarball, binaries_dir=p.strpath, buildinfo={"metadata": "new"}
     )
+
+    expected_prefix = p.strpath.lstrip("/")
 
     # Verify we don't have a repeated binary_distribution file,
     # and that the tarball contains the new one, not the old one.
     with tarfile.open(tarball) as tar:
-        assert syaml.load(tar.extractfile("my-pkg-prefix/.spack/binary_distribution")) == {
+        assert syaml.load(tar.extractfile(f"{expected_prefix}/.spack/binary_distribution")) == {
             "metadata": "new"
         }
         assert tar.getnames() == [
-            "my-pkg-prefix",
-            "my-pkg-prefix/.spack",
-            "my-pkg-prefix/.spack/binary_distribution",
+            f"{expected_prefix}",
+            f"{expected_prefix}/.spack",
+            f"{expected_prefix}/.spack/binary_distribution",
         ]
 
 
@@ -935,14 +934,16 @@ def test_reproducible_tarball_is_reproducible(tmpdir):
 
     # Create a tarball with a certain mtime of bin/app
     os.utime(app, times=(0, 0))
-    bindist._do_create_tarball(tarball_1, binaries_dir=p, pkg_dir="pkg", buildinfo=buildinfo)
+    bindist._do_create_tarball(tarball_1, binaries_dir=p.strpath, buildinfo=buildinfo)
 
     # Do it another time with different mtime of bin/app
     os.utime(app, times=(10, 10))
-    bindist._do_create_tarball(tarball_2, binaries_dir=p, pkg_dir="pkg", buildinfo=buildinfo)
+    bindist._do_create_tarball(tarball_2, binaries_dir=p.strpath, buildinfo=buildinfo)
 
     # They should be bitwise identical:
     assert filecmp.cmp(tarball_1, tarball_2, shallow=False)
+
+    expected_prefix = p.strpath.lstrip("/")
 
     # Sanity check for contents:
     with tarfile.open(tarball_1, mode="r") as f:
@@ -951,11 +952,11 @@ def test_reproducible_tarball_is_reproducible(tmpdir):
             assert m.uname == m.gname == ""
 
         assert set(f.getnames()) == {
-            "pkg",
-            "pkg/bin",
-            "pkg/bin/app",
-            "pkg/.spack",
-            "pkg/.spack/binary_distribution",
+            f"{expected_prefix}",
+            f"{expected_prefix}/bin",
+            f"{expected_prefix}/bin/app",
+            f"{expected_prefix}/.spack",
+            f"{expected_prefix}/.spack/binary_distribution",
         }
 
 
@@ -979,21 +980,23 @@ def test_tarball_normalized_permissions(tmpdir):
     with open(data, "w", opener=lambda path, flags: os.open(path, flags, 0o477)) as f:
         f.write("hello world")
 
-    bindist._do_create_tarball(tarball, binaries_dir=p, pkg_dir="pkg", buildinfo={})
+    bindist._do_create_tarball(tarball, binaries_dir=p.strpath, buildinfo={})
+
+    expected_prefix = p.strpath.lstrip("/")
 
     with tarfile.open(tarball) as tar:
         path_to_member = {member.name: member for member in tar.getmembers()}
 
     # directories should have 0o755
-    assert path_to_member["pkg"].mode == 0o755
-    assert path_to_member["pkg/bin"].mode == 0o755
-    assert path_to_member["pkg/.spack"].mode == 0o755
+    assert path_to_member[f"{expected_prefix}"].mode == 0o755
+    assert path_to_member[f"{expected_prefix}/bin"].mode == 0o755
+    assert path_to_member[f"{expected_prefix}/.spack"].mode == 0o755
 
     # executable-by-user files should be 0o755
-    assert path_to_member["pkg/bin/app"].mode == 0o755
+    assert path_to_member[f"{expected_prefix}/bin/app"].mode == 0o755
 
     # not-executable-by-user files should be 0o644
-    assert path_to_member["pkg/share/file"].mode == 0o644
+    assert path_to_member[f"{expected_prefix}/share/file"].mode == 0o644
 
 
 def test_tarball_common_prefix(dummy_prefix, tmpdir):
@@ -1062,3 +1065,50 @@ def test_tarfile_with_files_outside_common_prefix(tmpdir, dummy_prefix):
             ValueError, match="Tarball contains file /etc/config_file outside of prefix"
         ):
             bindist._ensure_common_prefix(tarfile.open("broken.tar", mode="r"))
+
+
+def test_tarfile_of_spec_prefix(tmpdir):
+    """Tests whether hardlinks, symlinks, files and dirs are added correctly,
+    and that the order of entries is correct."""
+    prefix = tmpdir.mkdir("prefix")
+    prefix.ensure("a_directory", dir=True).join("file").write("hello")
+    prefix.ensure("c_directory", dir=True).join("file").write("hello")
+    prefix.ensure("b_directory", dir=True).join("file").write("hello")
+    prefix.join("file").write("hello")
+    os.symlink(prefix.join("file"), prefix.join("symlink"))
+    os.link(prefix.join("file"), prefix.join("hardlink"))
+
+    file = tmpdir.join("example.tar")
+
+    with tarfile.open(file, mode="w") as tar:
+        bindist.tarfile_of_spec_prefix(tar, prefix.strpath)
+
+    expected_prefix = prefix.strpath.lstrip("/")
+
+    with tarfile.open(file, mode="r") as tar:
+        # Verify that entries are added in depth-first pre-order, files preceding dirs,
+        # entries ordered alphabetically
+        assert tar.getnames() == [
+            f"{expected_prefix}",
+            f"{expected_prefix}/file",
+            f"{expected_prefix}/hardlink",
+            f"{expected_prefix}/symlink",
+            f"{expected_prefix}/a_directory",
+            f"{expected_prefix}/a_directory/file",
+            f"{expected_prefix}/b_directory",
+            f"{expected_prefix}/b_directory/file",
+            f"{expected_prefix}/c_directory",
+            f"{expected_prefix}/c_directory/file",
+        ]
+
+        # Check that the types are all correct
+        assert tar.getmember(f"{expected_prefix}").isdir()
+        assert tar.getmember(f"{expected_prefix}/file").isreg()
+        assert tar.getmember(f"{expected_prefix}/hardlink").islnk()
+        assert tar.getmember(f"{expected_prefix}/symlink").issym()
+        assert tar.getmember(f"{expected_prefix}/a_directory").isdir()
+        assert tar.getmember(f"{expected_prefix}/a_directory/file").isreg()
+        assert tar.getmember(f"{expected_prefix}/b_directory").isdir()
+        assert tar.getmember(f"{expected_prefix}/b_directory/file").isreg()
+        assert tar.getmember(f"{expected_prefix}/c_directory").isdir()
+        assert tar.getmember(f"{expected_prefix}/c_directory/file").isreg()