Two changes in this PR:

1. Register absolute paths in tarballs*, which makes it easier
   to use them as container image layers, or rootfs in general, outside
   of Spack where no relocation can be done. Spack [supports installing](https://github.com/spack/spack/pull/37441)
   tarballs of this type already on develop. Tar file size would be
   a bit larger, but those repeated prefixes compress well. In practice
   I'm getting 0.4% size difference, so negligible.
2. Assemble the tarfile entries "by hand", which has a few advantages:
   1. Avoid reading `/etc/passwd`, `/etc/groups`, `/etc/nsswitch.conf`
      which `tar.add(dir)` does _for each file it adds_
   2. Reduce the number of stat calls per file added by a factor two,
      compared to `tar.add`, which should help with slow, shared filesystems
      where these calls are expensive
   4. Create normalized `TarInfo` entries from the start, instead of letting
      Python create them and patching them after the fact
   5. Don't recurse into subdirs before processing files, to avoid
      keeping nested directories opened. This changes the tar entry
      order slightly, it's like sorting by `(is_dir, name)` instead of `name`

Part 1 is necessary for #38358 

This change is forward incompatible, so old Spack 0.20 won't be able to
use tarballs generated with 0.21. (Unless #37441 is backported to 0.20)

\* the leading `/` is stripped, following GNU tar defaults, so technically the
paths are relative to the root `/` and not absolute.

### Creating a tarball from a python@3.11 install

Measured using

```
PYTHONPATH=$spack/lib/spack/external:$spack/lib/spack/external/_vendoring:$spack/lib/spack strace -fc python3 -c 'from spack.binary_distribution import _do_create_tarball; _do_create_tarball("a.tar.gz", "/path/to/python-3.11.4-ygq322p7xh65pm5suad2fzx6q7eyf7xs", "/path/to/python-3.11.4-ygq322p7xh65pm5suad2fzx6q7eyf7xs", {})'
```

on an NVMe disk, so probably multiply the seconds column by a factor 100-1000 to translate to HPC shared filesystem performance ;) 

Before this PR

```
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ------------------
 24.60    0.033036           0     52882       218 newfstatat
 18.32    0.024602           0     25669       270 openat
 17.97    0.024129           0     40214           read
 14.18    0.019037           0     25655         4 close
  8.64    0.011603           0     25198         9 lseek
  5.23    0.007017           2      2792           write
  1.04    0.001397           1       734           getdents64
```

After this

```
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ------------------
 25.39    0.013535           0     23634           read
 23.39    0.012466           0     19405       219 newfstatat
 13.14    0.007003           0      9090       270 openat
 11.37    0.006060           2      2792           write
  8.02    0.004275           0      8908         4 close
  6.58    0.003507           0      8620         9 lseek
  2.23    0.001186           1       734           getdents64
```

This includes (common) startup of Python itself... even then 2.7x fewer stat calls, and `read` dominating the syscalls as it should. `write` is of course small because of compression.