@alalazo 2023-10-02
In terms of wall time, on an HPC cluster, develop shows:
```console
$ time spack buildcache create ~/tmp/develop libpng
==> Selected 17 specs to push to file:///g/g92/culpo1/tmp/develop
gpg: checking the trustdb
gpg: marginals needed: 3  completes needed: 1  trust model: pgp
gpg: depth: 0  valid:   1  signed:   0  trust: 0-, 0q, 0n, 0m, 0f, 1u
==> [ 1/17] Pushed libpng@1.6.39/rrjrgaw
==> [ 2/17] Pushed cmake@3.27.6/mjqcazt
==> [ 3/17] Pushed curl@8.1.2/hsuxk52
==> [ 4/17] Pushed nghttp2@1.52.0/upwiaol
==> [ 5/17] Pushed pkgconf@1.9.5/itbpeol
==> [ 6/17] Pushed openssl@3.1.3/d542ymy
==> [ 7/17] Pushed ca-certificates-mozilla@2023-05-30/kuxdrfi
==> [ 8/17] Pushed perl@5.38.0/kvputle
==> [ 9/17] Pushed berkeley-db@18.1.40/oc4h6k6
==> [10/17] Pushed bzip2@1.0.8/64ouqrb
==> [11/17] Pushed diffutils@3.9/zh67wnb
==> [12/17] Pushed libiconv@1.17/b5olrx4
==> [13/17] Pushed gdbm@1.23/xcqqjcq
==> [14/17] Pushed readline@8.2/jns5mse
==> [15/17] Pushed ncurses@6.4/skzoqbs
==> [16/17] Pushed zlib-ng@2.1.3/qyygqi4
==> [17/17] Pushed gmake@4.4.1/kduz2h6

real	1m35,619s
user	1m3,254s
sys	0m15,350s
```
while this PR shows:
```console
$ time spack buildcache create ~/tmp/pr libpng
==> Selected 17 specs to push to file:///g/g92/culpo1/tmp/pr
==> [ 1/17] Pushed libpng@1.6.39/rrjrgaw
==> [ 2/17] Pushed cmake@3.27.6/mjqcazt
==> [ 3/17] Pushed curl@8.1.2/hsuxk52
==> [ 4/17] Pushed nghttp2@1.52.0/upwiaol
==> [ 5/17] Pushed pkgconf@1.9.5/itbpeol
==> [ 6/17] Pushed openssl@3.1.3/d542ymy
==> [ 7/17] Pushed ca-certificates-mozilla@2023-05-30/kuxdrfi
==> [ 8/17] Pushed perl@5.38.0/kvputle
==> [ 9/17] Pushed berkeley-db@18.1.40/oc4h6k6
==> [10/17] Pushed bzip2@1.0.8/64ouqrb
==> [11/17] Pushed diffutils@3.9/zh67wnb
==> [12/17] Pushed libiconv@1.17/b5olrx4
==> [13/17] Pushed gdbm@1.23/xcqqjcq
==> [14/17] Pushed readline@8.2/jns5mse
==> [15/17] Pushed ncurses@6.4/skzoqbs
==> [16/17] Pushed zlib-ng@2.1.3/qyygqi4
==> [17/17] Pushed gmake@4.4.1/kduz2h6

real	1m21,050s
user	0m53,942s
sys	0m13,770s
```
so a `17%` speedup using this PR.